Use separate files to force ViTest to behave
